+++ src/main/java/io/xlorey/FluxLoader/client/core/Core.java
@@ -9,6 +9,11 @@ import io.xlorey.FluxLoader.utils.Logger;
  */
 public class Core {
     /**
+     * Flag indicating whether client plugins have been loaded
+     */
+    private static boolean isClientLoaded = false;
+
+    /**
      * Initializing the loader
      * @exception Exception in cases of unsuccessful core initialization
      */
@@ -19,8 +24,21 @@ public class Core {
 
         EventManager.subscribe(new EventsHandler());
 
-        PluginManager.loadPlugins(true);
+        PluginManager.loadPlugins();
 
         WidgetManager.init();
+
+        isClientLoaded = true;
+    }
+
+    /**
+     * Checks if client plugins have been loaded.
+     * This method provides information about the initialization state of the FluxLoader client.
+     * It returns the value of a flag that indicates whether initialization was successful
+     * and loading client plugins.
+     * @return true if client plugins were loaded successfully; otherwise false.
+     */
+    public static boolean isClientLoaded() {
+        return isClientLoaded;
     }
 }