+++ src/main/java/io/xlorey/FluxLoader/server/core/EventsHandler.java
@@ -1,12 +1,7 @@
 package io.xlorey.FluxLoader.server.core;
 
 import io.xlorey.FluxLoader.annotations.SubscribeEvent;
-import io.xlorey.FluxLoader.plugin.Plugin;
 import io.xlorey.FluxLoader.shared.PluginManager;
-import io.xlorey.FluxLoader.utils.Logger;
-
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * Server events handler
@@ -17,21 +12,6 @@ public class EventsHandler {
      */
     @SubscribeEvent(eventName="onServerShutdown")
     public void onServerShutdownHandler(){
-        HashMap<String, Plugin> pluginsRegistry = PluginManager.getLoadedPlugins();
-
-        for (Map.Entry<String, Plugin> entry : pluginsRegistry.entrySet()) {
-            String pluginId = entry.getKey();
-            Plugin pluginInstance = entry.getValue();
-
-            Logger.print(String.format("Plugin '%s' is starting to shut down...", pluginId));
-
-            try {
-                pluginInstance.onTerminate();
-            } catch (Exception e) {
-                Logger.print(String.format("Plugin '%s' failed to shut down correctly due to: %s", pluginId, e));
-            }
-
-            Logger.print(String.format("Plugin '%s' has been successfully terminated.", pluginId));
-        }
+        PluginManager.terminatePlugins(PluginManager.getLoadedServerPlugins());
     }
 }